+++ spec/details/server.lua
@@ -55,7 +55,7 @@ local function tcp_server(port, opts)
         end
 
         if opts.tls and handshake_done then
-          local ssl = require "spec.helpers.ssl"
+          local ssl = require "spec.details.ssl"
 
           local params = {
             mode = "server",
